+++ tests/test_phase_psf.py
@@ -197,7 +197,7 @@ def test_frozen_flow():
         aper = galsim.Aperture(diam=1, pupil_plane_size=20., pupil_plane_scale=20./dx)
     wf0 = screen.wavefront(aper)
     screen.advance_by(t)
-    wf1 = screen.wavefront(aper, theta_x=45*galsim.degrees)
+    wf1 = screen.wavefront(aper, theta=(45*galsim.degrees, 0*galsim.degrees))
 
     np.testing.assert_array_almost_equal(wf0, wf1, 5, "Flow is not frozen")
 
@@ -241,8 +241,7 @@ def test_phase_psf_batch():
     exptime = 0.06
     rng = galsim.BaseDeviate(1234)
     atm = galsim.Atmosphere(screen_size=10.0, altitude=10.0, alpha=0.997, rng=rng)
-    theta_x = [i * galsim.arcsec for i in range(NPSFs)]
-    theta_y = [i * galsim.arcsec for i in range(NPSFs)]
+    theta = [(i*galsim.arcsec, i*galsim.arcsec) for i in range(NPSFs)]
 
     kwargs = dict(lam=500.0, exptime=exptime, diam=4.0,
                   pupil_plane_size=6.0, pupil_plane_scale=6.0/192)
@@ -251,35 +250,17 @@ def test_phase_psf_batch():
     import warnings
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
-        psfs = atm.makePSF(theta_x=theta_x, theta_y=theta_y, **kwargs)
+        psfs = atm.makePSF(theta=theta, **kwargs)
     print('time for {0} PSFs in batch: {1:.2f} s'.format(NPSFs, time.time() - t1))
 
     t2 = time.time()
     more_psfs = []
-    for tx, ty in zip(theta_x, theta_y):
+    for th in theta:
         atm.reset()
-        more_psfs.append(atm.makePSF(theta_x=tx, theta_y=ty, **kwargs))
+        more_psfs.append(atm.makePSF(theta=th, **kwargs))
     print('time for {0} PSFs in serial: {1:.2f} s'.format(NPSFs, time.time() - t2))
 
     for psf1, psf2 in zip(psfs, more_psfs):
-        np.testing.assert_array_equal(
-            psf1.img, psf2.img,
-            "Individually generated AtmosphericPSF differs from AtmosphericPSF generated in batch")
-
-    # Check that `theta` kwarg works.
-    atm.reset()
-    even_more_psfs = atm.makePSF(theta=zip(theta_x, theta_y), **kwargs)
-    for psf1, psf2 in zip(psfs, even_more_psfs):
-        np.testing.assert_array_equal(
-            psf1.img, psf2.img,
-            "Individually generated AtmosphericPSF differs from AtmosphericPSF generated in batch")
-
-    # Check that `theta` kwarg works 1 PSF at a time.
-    yet_another_list_of_psfs = []
-    for th in zip(theta_x, theta_y):
-        atm.reset()
-        yet_another_list_of_psfs.append(atm.makePSF(theta=th, **kwargs))
-    for psf1, psf2 in zip(psfs, yet_another_list_of_psfs):
         np.testing.assert_array_equal(
             psf1.img, psf2.img,
             "Individually generated AtmosphericPSF differs from AtmosphericPSF generated in batch")